+++ yaralyzer/__init__.py
@@ -24,6 +24,7 @@ PDFALYZER_MSG_TXT.append('https://github.com/michelcrypt4d4mus/pdfalyzer\n', sty
 
 
 def yaralyze():
+    """Main entry point for yaralyzer when invoked as a script. Args are parsed from the command line."""
     args = parse_arguments()
     output_basepath = None